% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{elnet_fun}
\alias{elnet_fun}
\title{Wrapper for glmnet at a specific lambda}
\usage{
elnet_fun(lambda, alpha, family = "gaussian")
}
\arguments{
\item{lambda}{What lambda to use in the optimization?}
}
\value{
A function that fits the elastic net given only two arguments (x and y)
at the specified lambda.
}
\description{
Wrapper for glmnet at a specific lambda
}

