% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{independent_models}
\alias{independent_models}
\title{Fit an independent model for each response coordinate}
\usage{
independent_models(X, Y, model_fun)
}
\arguments{
\item{X}{The N x P covariates matrix. NA values will be dropped.}

\item{Y}{an N x J response matrix}

\item{model_fun}{A function that returns coefficients Y[, j] ~ X for each j.}
}
\value{
The J X P coefficients matrix.
}
\description{
Fit an independent model for each response coordinate
}

