% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lf_logit_spline.R
\name{lf_spline_logit}
\alias{lf_spline_logit}
\title{Logistic regression with spline basis}
\usage{
lf_spline_logit(Y, H, B0, W0, opts = list())
}
\description{
Logistic regression with spline basis
}

