% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/latent_factor.R
\name{lsam_objective}
\alias{lsam_objective}
\title{Compute objective for alternating minimization}
\usage{
lsam_objective(Y, param)
}
\arguments{
\item{Y}{The data to which we aim to fit.}

\item{param}{A list with the elements Phi and W, for the source and
coefficients in the latent source model.}
}
\value{
The RSS for the difference, and the l1 + l2 norms of the coefficients
}
\description{
Compute objective for alternating minimization
}

