% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline_lf.R
\name{merge_spline_lf_opts}
\alias{merge_spline_lf_opts}
\title{Merge in default options for spline_lf()}
\usage{
merge_spline_lf_opts(opts = list())
}
\arguments{
\item{opts}{A potentially partially specified list, which we will merge in
with defaults. Please include new parameters in alphabetical order. The
possible options, and their defaults are, \cr
 $alpha The tradeoff between l1 and l2 penalties in the elastic net. \cr
 $lambda The overall regularization parameter in the elastic net. \cr
 $nu The learning rate / step-size for the coordinate descent.
  Arbitrarily defaults to 1e-5. If things are not converging, try making this
  smaller. \cr
 $n_iter The number of steps of the alternating minimization to perform. \cr
 $n_iter_B The number of sweeps over all entries of B via coordinate descent.}
}
\value{
The modified opts list, with defaults filled in for entries that had
been missing.
}
\description{
Merge in default options for spline_lf()
}

