% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spline_lf.R
\name{spline_lf_objective}
\alias{spline_lf_objective}
\title{Compute terms in the objective for the spline latent factor model}
\usage{
spline_lf_objective(Y, H, param)
}
\description{
Compute terms in the objective for the spline latent factor model
}

